[REFACTOR] Move select element to common files
