+++ pandora-client-web/src/components/settings/helpers/settingsInputs.tsx
@@ -2,9 +2,9 @@ import { EMPTY_ARRAY, KnownObject } from 'pandora-common';
 import React, { useCallback, useId, useMemo, type DependencyList, type ReactElement, type ReactNode } from 'react';
 import type { ZodSchema, ZodTypeDef } from 'zod';
 import { useRemotelyUpdatedUserInput } from '../../../common/useRemotelyUpdatedUserInput';
+import { Select, type SelectProps } from '../../../common/userInteraction/select/select';
 import { Button } from '../../common/button/button';
 import { Row } from '../../common/container/container';
-import { Select, type SelectProps } from '../../common/select/select';
 
 export function ToggleSettingInput({ currentValue, defaultValue, label, onChange, onReset, deps = EMPTY_ARRAY }: {
 	currentValue: boolean | undefined;
