+++ pandora-client-web/src/components/wardrobe/wardrobeItemPreferences.tsx
@@ -8,10 +8,10 @@ import { z } from 'zod';
 import { useAssetManager } from '../../assets/assetManager';
 import { useBrowserStorage } from '../../browserStorage';
 import { useInputAutofocus } from '../../common/userInteraction/inputAutofocus';
+import { Select, type SelectProps } from '../../common/userInteraction/select/select';
 import { TOAST_OPTIONS_ERROR } from '../../persistentToast';
 import { Column, Row } from '../common/container/container';
 import { Scrollable, Scrollbar } from '../common/scrollbar/scrollbar';
-import { Select, type SelectProps } from '../common/select/select';
 import { Tab, TabContainer } from '../common/tabs/tabs';
 import { useShardConnector } from '../gameContext/shardConnectorContextProvider';
 import { WardrobeAssetList, useAssetPreference, useAssetPreferenceResolver, useAssetPreferences } from './views/wardrobeAssetView';
