+++ pandora-client-web/src/editor/editor.tsx
@@ -9,8 +9,8 @@ import { GetCurrentAssetManager } from '../assets/assetManager';
 import { GraphicsManager } from '../assets/graphicsManager';
 import { useBrowserStorage } from '../browserStorage';
 import { useEvent } from '../common/useEvent';
+import { Select } from '../common/userInteraction/select/select';
 import { Button } from '../components/common/button/button';
-import { Select } from '../components/common/select/select';
 import { LocalErrorBoundary } from '../components/error/localErrorBoundary';
 import { LayerStateOverrides } from '../graphics/def';
 import { Observable } from '../observable';
