+++ pandora-client-web/src/ui/components/chat/chatInput.tsx
@@ -10,11 +10,11 @@ import { BrowserStorage } from '../../../browserStorage';
 import { Character } from '../../../character/character';
 import { useEvent } from '../../../common/useEvent';
 import { useInputAutofocus } from '../../../common/userInteraction/inputAutofocus';
+import { Select, type SelectProps } from '../../../common/userInteraction/select/select';
 import { useTextFormattingOnKeyboardEvent } from '../../../common/useTextFormattingOnKeyboardEvent';
 import { Button } from '../../../components/common/button/button';
 import { Column, Row } from '../../../components/common/container/container';
 import { Scrollable } from '../../../components/common/scrollbar/scrollbar';
-import { Select, type SelectProps } from '../../../components/common/select/select';
 import { useDirectoryConnector } from '../../../components/gameContext/directoryConnectorContextProvider';
 import { IMessageParseOptions, useChatCharacterStatus, useChatMessageSender, useChatSetPlayerStatus, useGameState, useGameStateOptional, useSpaceCharacters } from '../../../components/gameContext/gameStateContextProvider';
 import { usePlayerId } from '../../../components/gameContext/playerContextProvider';
