+++ pandora-client-web/src/ui/screens/spaceConfiguration/spaceConfiguration.tsx
@@ -37,12 +37,12 @@ import { CopyToClipboard } from '../../../common/clipboard';
 import { useCurrentTime } from '../../../common/useCurrentTime';
 import { useAsyncEvent } from '../../../common/useEvent';
 import { useInputAutofocus } from '../../../common/userInteraction/inputAutofocus';
+import { Select } from '../../../common/userInteraction/select/select';
 import { Button } from '../../../components/common/button/button';
 import { ColorInput } from '../../../components/common/colorInput/colorInput';
 import { Column, Row } from '../../../components/common/container/container';
 import { FieldsetToggle } from '../../../components/common/fieldsetToggle';
 import { Scrollbar } from '../../../components/common/scrollbar/scrollbar';
-import { Select } from '../../../components/common/select/select';
 import { SelectionIndicator } from '../../../components/common/selectionIndicator/selectionIndicator';
 import { Tab, TabContainer } from '../../../components/common/tabs/tabs';
 import { ModalDialog, useConfirmDialog } from '../../../components/dialog/dialog';
